requirements_dev.txt
jupyterlab==3.6.3
matplotlib==3.7.1
numpy>=1.26.4, <2.0.0
pandas==2.0.1
scikit-learn==1.2.2
statsmodels==0.14.0
pytest==7.4.0
testbook==0.4.2
parsenvy==3.0.2
black==23.3.0
protobuf==4.23.3
nbdime==4.0.2
python-dotenv==1.0.1
langchain-community==0.3.29
langchain-openai==0.3.32
langchain-huggingface==0.3.1
sentence-transformers==5.1.0
langchain-groq==0.3.8
langchainhub==0.1.21
pypdf==6.0.0
faiss-cpu==1.12.0
plotly==6.3.0
supabase==1.0.0